Now more compatible with Ubuntu 12.04 & new features
- Fixes: Made the script more compatible with Ubuntu 12.04 LTS
(Precise Pangolin) using the ports.conf file, pinky instead of finger,
checks apache2.conf for include statement of the Ubuntu/Debian standard
'sites-enabled' configuration directory and add it if it's not found.

- New: Added command line parameter '--version' to check for the version
of the script.

- New: Added CREATE_INDEX variable. Set to yes to explicitly create an
index.html file if none was found. By default set to no. So no index.html
will be created from now on.

- New: Added ERROR_LOG variable. Set to /var/log/apache2, but can be easily
easily changed to set the VirtualHost's errorlog. Uses the following format:
$VIRTUALHOST-error.log